+++ packages/nextjs/src/server.ts
@@ -2,4 +2,11 @@
 // client modules because otherwise the consumer's next bundler
 // will include client modules in server code
 export { createNextRouteHandler } from './createNextRouteHandler'
-export { FlowgladServer } from '@flowglad/server'
+export {
+  FlowgladServer,
+  createRequestHandler,
+  RequestHandlerError,
+  RequestHandlerOptions,
+  RequestHandlerInput,
+  RequestHandlerOutput,
+} from '@flowglad/server'