+++ packages/server/src/index.ts
@@ -1,2 +1,11 @@
 export { FlowgladServer } from './flowgladServer'
 export { routeToHandlerMap } from './subrouteHandlers'
+export {
+  createRequestHandler,
+  RequestHandlerError,
+} from './requestHandler'
+export type {
+  RequestHandlerInput,
+  RequestHandlerOutput,
+  RequestHandlerOptions,
+} from './requestHandler'